# ProyectoFinal/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte Final ",
        "description": "REST API. Projecte Final. DAW Client i servidor Joan R, Dimitry C, Javier P.",
        "contact": {
            "name": "Joan Ramis.",
            "email": "joanramis1@paucasesnovescifp.cat"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://www.rampacom.com/ProyectoFinal/public/"
        },
        {
            "url": "http://127.0.0.1/projecte/TreballFinal/ProyectoFinal/public/"
        }
    ],
    "paths": {
...
        "/api/idioma/{id}": {
            "get": {
                "tags": [
                    "Idiomas"
                ],
                "summary": "Mostrar un Idioma por ID",
                "description": "Descripcion de un Idioma.",
                "operationId": "709f85a3b3076471b218f91fbe1ed831",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id del Idioma",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "id": {
                                "summary": "Introduce el numero de ID del Idioma",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informacion del Idioma.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Hay un error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Idioma no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/idioma/borra/{id}": {
            "delete": {
                "tags": [
                    "Idiomas"
                ],
                "summary": "Borra un Idioma",
                "description": "Borra un Idioma. Solo por Administradores",
                "operationId": "52df9fcec717a89c43a5604ea62cb6a6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Idioma",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "ID no encotrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/idioma/crea": {
            "post": {
                "tags": [
                    "Idiomas"
                ],
                "summary": "Crea un Idioma",
                "description": "Crea un Idioma. Solo por Administradores.",
                "operationId": "329cc4285399793eebaebfbad6bfc5f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idioma": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Esto es un nuevo Idioma"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Atributo obligatorio requerido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/idioma/modifica/{id}": {
            "put": {
                "tags": [
                    "Idiomas"
                ],
                "summary": "Modifica un Idioma",
                "description": "Modifica un Idioma. Solo por Administradores.",
                "operationId": "4c145f4b3bb369d58faa84696925cd57",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Idioma",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "idioma": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Esto es un nuevo Idioma"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Atributo obligatorio requerido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}